requirements.txt
httpx
motor
matplotlib
flake8
black
python-dotenv